GH-586 Move snapshotted_tester to header file for reuse in partitioned_block_log_tests
